Add commands-roleplay into core-commands.
