+++ scripts/buildresource/resource.json
@@ -6,7 +6,6 @@
 		"client/*",
 		"shared/*",
 		"plugins/athena/client/*",
-		"plugins/commands-moderator/client/*",
 		"plugins/commands-roleplay/client/*",
 		"plugins/commands-roleplay/shared/*",
 		"plugins/core-atm/client/*",
@@ -17,6 +16,8 @@
 		"plugins/core-character-select/shared/*",
 		"plugins/core-clothing/client/*",
 		"plugins/core-clothing/shared/*",
+		"plugins/core-commands/client/*",
+		"plugins/core-commands/shared/*",
 		"plugins/core-console-cmds/client/*",
 		"plugins/core-console-cmds/shared/*",
 		"plugins/core-cuffs/client/*",
