+++ src/core/plugins/core-commands/server/config/index.ts
