+++ src/core/plugins/core-commands/shared/iRoleplayCmds.ts
